requirements.txt
requests
numpy
pandas
psycopg2-binary
python-dotenv
openpyxl
xlrd 
sqlalchemy